// src/shared/security/utils/role.enum.ts
export enum Role {
    ADMIN_RESENIA = "adminResenia",
    ADMIN_PRODUCTO = "adminProducto",
    ADMIN_PROMOCION = "adminPromocion",
    ADMIN_MENU = "adminMenu",
    ADMIN_RESERVA = "adminReserva",
    ADMIN_USUARIO = "adminUsuario",
    ADMIN_CLIENTE = "adminCliente",
    ADMIN_ADMINISTRADOR_ESTABLECIMIENTO = "adminAdministradorEstablecimiento",
    ADMIN_ESTABLECIMIENTO = "adminEstablecimiento",
    ADMIN_TAG = "adminTag",
    READ_RESENIA = "readResenia",
    READ_PRODUCTO = "readProducto",
    READ_PROMOCION = "readPromocion",
    READ_MENU = "readMenu",
    READ_RESERVA = "readReserva",
    READ_USUARIO = "readUsuario",
    READ_CLIENTE = "readCliente",
    READ_ADMINISTRADOR_ESTABLECIMIENTO = "readAdministradorEstablecimiento",
    READ_ESTABLECIMIENTO = "readEstablecimiento",
    READ_TAG = "readTag",
    WRITE_RESENIA = "writeResenia",
    WRITE_PRODUCTO = "writeProducto",
    WRITE_PROMOCION = "writePromocion",
    WRITE_MENU = "writeMenu",
    WRITE_RESERVA = "writeReserva",
    WRITE_USUARIO = "writeUsuario",
    WRITE_CLIENTE = "writeCliente",
    WRITE_ADMINISTRADOR_ESTABLECIMIENTO = "writeAdministradorEstablecimiento",
    WRITE_ESTABLECIMIENTO = "writeEstablecimiento",
    WRITE_TAG = "writeTag",
    DELETE_RESENIA = "deleteResenia",
    DELETE_PRODUCTO = "deleteProducto",
    DELETE_PROMOCION = "deletePromocion",
    DELETE_MENU = "deleteMenu",
    DELETE_RESERVA = "deleteReserva",
    DELETE_USUARIO = "deleteUsuario",
    DELETE_CLIENTE = "deleteCliente",
    DELETE_ADMINISTRADOR_ESTABLECIMIENTO = "deleteAdministradorEstablecimiento",
    DELETE_ESTABLECIMIENTO = "deleteEstablecimiento",
    DELETE_TAG = "deleteTag",
}